Code review fixes: correctness, concurrency, perf; remove OpenVMS port
Full review findings in 00-claude-code-review.md, work log in
10-claude-work-done.md. Highlights:

- Linux: read /proc/<pid>/net/* (target netns) instead of /proc/net/*;
  socket inodes as uint64_t (were silently dropped above 2^31); stricter
  /proc stat validation; per-library RSS from smaps
- Model: MemoryMapInfo.size_bytes and LibraryInfo.base_addr numeric fields
  (kills Solaris display-string size round-trip and broken hex-string sort)
- FreeBSD: real executable path via KERN_PROC_PATHNAME instead of argv[0]
- Cross-platform consistency: priority semantics, is_executable = main
  binary, error retention (10 entries / 10 s) on all backends
- DataStore: fix lost-wakeup race in stop/refresh_now/resume/interval;
  guard CPU counter regression underflow
- SingleInstance: join listener before closing fd (teardown race)
- TUI: render only on input/data/resize with timeout(50) getch instead of
  60 fps busy redraw loop
- Search (GUI+TUI): match command line, search collapsed subtrees and
  auto-expand ancestors of matches
- Remove abandoned OpenVMS port (sources, DESCRIP.MMS, tools, all __VMS
  ifdefs in shared code)

Verified: clean Release build of pex + pexc on Debian 13 (zero warnings),
pexc smoke-tested in a pty.

+++ src/platform/linux/procfs_reader.hpp
@@ -41,7 +41,7 @@ private:
     std::string get_username(int uid);
     std::map<int, std::string> uid_cache_;
 
-    static std::map<int, NetworkConnectionInfo> parse_net_file(const std::string& path, const std::string& protocol);
+    static std::map<uint64_t, NetworkConnectionInfo> parse_net_file(const std::string& path, const std::string& protocol);
 
     // Error tracking
     void add_error(const std::string& message);
